refactor: apply clippy auto-fix for code cleanup
- Simplify iterator patterns with filter_map and flatten
- Use more idiomatic Rust patterns (if let, matches!)
- Remove redundant closures and type annotations
- Clean up string handling and formatting
- Improve Option/Result chain ergonomics

Auto-fixed 30 files across sage-core, sage-cli, and sage-tools.

+++ crates/sage-core/src/llm/client.rs
@@ -167,7 +167,7 @@ impl LLMClient {
 
                     if attempt < max_retries {
                         // Calculate exponential backoff with jitter
-                        let base_delay_secs = 2_u64.pow(attempt as u32);
+                        let base_delay_secs = 2_u64.pow(attempt);
                         let jitter_ms = {
                             let mut rng = rand::thread_rng();
                             rng.gen_range(0..=(base_delay_secs * 500))
